### 2021-09-29 #3779 @whedon
[PRE REVIEW]: robustHD: An R package for robust regression with high-dimensional data
**Submitting author:** @aalfons (<a href="http://orcid.org/0000-0002-2513-3788">Andreas Alfons</a>)
**Repository:** <a href="https://github.com/aalfons/robustHD/" target ="_blank">https://github.com/aalfons/robustHD/</a>
**Version:** 0.7.0
