### 2021-09-29 c26649b
added scripts for JOSS manuscript examples
