### 2021-09-29 af8a6f0
fixed some references in JOSS manuscript
