### 2021-09-29 e09745c
updated DOI's for JOSS manuscript
