+++ inst/doc/paper/paper.bib
@@ -23,7 +23,7 @@
   number = {480},
   pages = {1289--1299},
   year = {2007},
-  url = {https://doi.org/10.1198/016214507000000950}
+  DOI = {10.1198/016214507000000950}
 }
 
 @article{alfons13b,
@@ -34,7 +34,7 @@
   number = {1},
   pages = {226--248},
   year = {2013},
-  url = {https://doi.org/10.1214/12-AOAS575}
+  DOI = {10.1214/12-AOAS575}
 }
 
 @article{alfons16a,
@@ -44,7 +44,7 @@
   volume = {93},
   pages = {421--435},
   year = {2016},
-  url = {https://doi.org/10.1016/j.csda.2015.02.007}
+  DOI = {10.1016/j.csda.2015.02.007}
 }
 
 @article{tibshirani96,
@@ -55,7 +55,7 @@
   Number = {1},
   Pages = {267--288},
   Year = {1996},
-  url = {https://doi.org/10.1111/j.2517-6161.1996.tb02080.x}
+  DOI = {10.1111/j.2517-6161.1996.tb02080.x}
 }
 
 @article{smucler17,
@@ -65,7 +65,7 @@
   Volume = {111},
   Pages = {116--130},
   Year = {2017},
-  url = {https://doi.org/10.1016/j.csda.2017.02.002}
+  DOI = {10.1016/j.csda.2017.02.002}
 }
 
 @article{kurnaz18,
@@ -75,7 +75,7 @@
   Volume = {172},
   Pages = {211--222},
   Year = {2018},
-  url = {https://doi.org/10.1016/j.chemolab.2017.11.017}
+  DOI = {10.1016/j.chemolab.2017.11.017}
 }
 
 @article{chang18,
@@ -86,7 +86,7 @@
   Number = {1},
   Pages = {36--47},
   Year = {2018},
-  url = {https://doi.org/10.1080/00401706.2017.1305299}
+  DOI = {10.1080/00401706.2017.1305299}
 }
 
 @article{cohenfreue19,
@@ -97,7 +97,7 @@
   number = {4},
   pages = {2065--2090},
   year = {2019},
-  url = {https://doi.org/10.1214/19-AOAS1269}
+  DOI = {10.1214/19-AOAS1269}
 }
 
 @article{stadlbauer20,
@@ -108,7 +108,7 @@
   Number = {248},
   Pages = {1--13},
   Year = {2020},
-  url = {https://doi.org/10.1186/s12877-020-01644-2}
+  DOI = {10.1186/s12877-020-01644-2}
 }
 
 @article{antczakorlewska21,
@@ -119,7 +119,7 @@
   Number = {583831},
   Pages = {1--16},
   Year = {2021},
-  url = {https://doi.org/10.3389/fevo.2020.583831}
+  DOI = {10.3389/fevo.2020.583831}
 }
 
 @article{reinhold12,
@@ -130,7 +130,7 @@
   Number = {14},
   Pages = {3499--3511},
   Year = {2012},
-  url = {https://doi.org/10.1158/0008-5472.can-12-1370}
+  DOI = {10.1158/0008-5472.can-12-1370}
 }
 
 @manual{enetLTS,
