### 2021-09-29 #3779 @aalfons
@danielskatz - My apologies, I added the DOI URL's instead of just the DOI's. I updated the bib file and I'll try again now to see if I did it correctly.
