### 2021-09-30 #3779 @aalfons
@mikldk - sure: **valentint** is one of the authors of the `robustbase` package, which is the main package on robust statistics in R, and he has also written the package `rrcovHD` which contains functionality such as PCA and outlier detection in high dimensions.
