### 2021-09-30 #3779 @mikldk
@aalfons. Thanks. A few comments before I invite the reviewers:

- [ ] Can you add your country to the paper (in addition to "Rrasmus School of Economics, Erasmus University Rotterdam")?
- [ ] Figure 1: Change `lambda` to `expression(lambda)`
- [ ] The paper is fairly long (see [here](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain)); maybe move some examples to a `README.Rmd` (that is then compiled to `README.md`)?
