### 2021-09-30 #3779 @aalfons
**Point 1)** Done.

**Point 2)** I deliberately used `lambda` in Figure 1 to maintain the connection with argument `lambda` of function `sparseLTS()`. However, if you prefer me to use `expression(lambda)`, I can change that.

**Point 3)** When I checked the word count in `paper.md` with RStudio before submission, I got 1017 words excluding the metadata, so I was hoping it would still be ok. I designed the two examples to show off most of the functionality of the package (regularization, stepwise variable selection along a sequence, model selection via prediction error and via information criterion, examples of plots) in a rather compact form (6 pages including references). But if you prefer, I can remove the second example (robust groupwise least angle regression) from the paper. In that case I would still put *both examples* in `README.Rmd`, to keep the set of examples (that shows off most functionality) together in one place.

@mikldk - please let me know what you think and I'll make the changes accordingly.
