### 2021-09-30 #3779 @mikldk
@aalfons Thanks. Ad 2: Fine with me :+1:. Ad 3: Let's hear what the reviewers say. Anyway, I think it would be nice to have the examples in the `README`, too.
