+++ README.md
@@ -1,70 +1,293 @@
-# robustHD: Robust Methods for High-Dimensional Data
 
+# robustHD: Robust Methods for High-Dimensional Data
 
 ## Summary
 
-In regression analysis with high-dimensional data, variable selection is an important step to (i) overcome computational problems, (ii) improve prediction performance by variance reduction, and (iii) increase interpretability of the resulting models due to the smaller number of variables.  However, robust methods are necessary to prevent outlying data points from distorting the results.  The add-on package `robustHD` for the statistical computing environment `R` provides functionality for robust linear model selection with high-dimensional data.  More specifically, the implemented functionality includes robust least angle regression ([Khan et al., 2007](https://doi.org/10.1198/016214507000000950)), robust groupwise least angle regression ([Alfons et al., 2016](https://doi.org/10.1016/j.csda.2015.02.007)), as well as sparse least trimmed squares regression [Alfons et al., 2013](https://doi.org/10.1214/12-AOAS575). The latter can be seen as a trimmed version of the popular lasso regression estimator ([Tibshirani, 1996](https://doi.org/10.1111/j.2517-6161.1996.tb02080.x)).  Selecting the optimal model can be done via cross-validation or an information criterion, and various plots are available to illustrate model selection and to evaluate the final model estimates.  Furthermore, the package includes functionality for pre-processing and cleaning the data, such as robust standardization and winsorization.  Finally, `robustHD` follows a clear object-oriented design and takes advantage of C++ code and parallel computing to reduce computing time. 
-
+In regression analysis with high-dimensional data, variable selection is
+an important step to (i) overcome computational problems, (ii) improve
+prediction performance by variance reduction, and (iii) increase
+interpretability of the resulting models due to the smaller number of
+variables. However, robust methods are necessary to prevent outlying
+data points from distorting the results. The add-on package `robustHD`
+for the statistical computing environment `R` provides functionality for
+robust linear model selection with high-dimensional data. More
+specifically, the implemented functionality includes robust least angle
+regression ([Khan et al.,
+2007](https://doi.org/10.1198/016214507000000950)), robust groupwise
+least angle regression ([Alfons et al.,
+2016](https://doi.org/10.1016/j.csda.2015.02.007)), as well as sparse
+least trimmed squares regression [Alfons et al.,
+2013](https://doi.org/10.1214/12-AOAS575). The latter can be seen as a
+trimmed version of the popular lasso regression estimator ([Tibshirani,
+1996](https://doi.org/10.1111/j.2517-6161.1996.tb02080.x)). Selecting
+the optimal model can be done via cross-validation or an information
+criterion, and various plots are available to illustrate model selection
+and to evaluate the final model estimates. Furthermore, the package
+includes functionality for pre-processing and cleaning the data, such as
+robust standardization and winsorization. Finally, `robustHD` follows a
+clear object-oriented design and takes advantage of C++ code and
+parallel computing to reduce computing time.
 
 ## Main functionality
 
- * `sparseLTS()`: Sparse least trimmed squares regression.
- 
- * `rlars()`: Robust least angle regression.
- 
- * `grplars()` and `rgrplars()`: (Robust) groupwise least angle regression.
- 
- * `tslars()` and `rtslars()`: (Robust) least angle regression for time series data.
- 
- * `corHuber()`: Robust correlation based on winsorization.
- 
- * `winsorize()`: Data cleaning by winsorization.
- 
[…]
