+++ README.md
@@ -166,7 +166,7 @@ regularization parameter. Moreover, function `diagnosticPlot()` allows
 to produce various diagnostic plots for the optimal model fit.
 
 <figure>
-<embed src="https://github.com/aalfons/robustHD/blob/master/inst/doc/paper/figure_sparseLTS.pdf?raw=true" style="width:67.0%" /><figcaption aria-hidden="true">Examples of the optimality criterion plot (<em>left</em>) and the regression diagnostic plot (<em>right</em>) for output of function <code>sparseLTS()</code> in package <code>robustHD</code>.</figcaption>
+<embed src="https://github.com/aalfons/robustHD/blob/master/inst/doc/paper/figure_sparseLTS.png?raw=true" style="width:67.0%" /><figcaption aria-hidden="true">Examples of the optimality criterion plot (<em>left</em>) and the regression diagnostic plot (<em>right</em>) for output of function <code>sparseLTS()</code> in package <code>robustHD</code>.</figcaption>
 </figure>
 
 # Example: Robust groupwise least angle regression
@@ -249,7 +249,7 @@ coefficient path along the sequence can be produced via the function
 available to produce similar plots as in the previous example.
 
 <figure>
-<embed src="https://github.com/aalfons/robustHD/blob/master/inst/doc/paper/figure_rgrplars.pdf?raw=true" style="width:100.0%" /><figcaption aria-hidden="true">Examples of the coefficient plot (<em>left</em>), the optimality criterion plot (<em>center</em>), and the regression diagnostic plot (<em>right</em>) for output of function <code>rgrplars()</code> in package <code>robustHD</code>.</figcaption>
+<embed src="https://github.com/aalfons/robustHD/blob/master/inst/doc/paper/figure_rgrplars.png?raw=true" style="width:100.0%" /><figcaption aria-hidden="true">Examples of the coefficient plot (<em>left</em>), the optimality criterion plot (<em>center</em>), and the regression diagnostic plot (<em>right</em>) for output of function <code>rgrplars()</code> in package <code>robustHD</code>.</figcaption>
 </figure>
 
 ## Community guidelines
