+++ README.md
@@ -166,7 +166,7 @@ regularization parameter. Moreover, function `diagnosticPlot()` allows
 to produce various diagnostic plots for the optimal model fit.
 
 <figure>
-<img src="./inst/doc/paper/figure_sparseLTS.png" style="width:67.0%" alt="Examples of the optimality criterion plot (left) and the regression diagnostic plot (right) for output of function sparseLTS() in package robustHD." /><figcaption aria-hidden="true">Examples of the optimality criterion plot (<em>left</em>) and the regression diagnostic plot (<em>right</em>) for output of function <code>sparseLTS()</code> in package <code>robustHD</code>.</figcaption>
+<img src="./inst/doc/paper/figure_sparseLTS.svg" style="width:67.0%" alt="Examples of the optimality criterion plot (left) and the regression diagnostic plot (right) for output of function sparseLTS() in package robustHD." /><figcaption aria-hidden="true">Examples of the optimality criterion plot (<em>left</em>) and the regression diagnostic plot (<em>right</em>) for output of function <code>sparseLTS()</code> in package <code>robustHD</code>.</figcaption>
 </figure>
 
 # Example: Robust groupwise least angle regression
@@ -249,7 +249,7 @@ coefficient path along the sequence can be produced via the function
 available to produce similar plots as in the previous example.
 
 <figure>
-<img src="./inst/doc/paper/figure_rgrplars.png" style="width:100.0%" alt="Examples of the coefficient plot (left), the optimality criterion plot (center), and the regression diagnostic plot (right) for output of function rgrplars() in package robustHD." /><figcaption aria-hidden="true">Examples of the coefficient plot (<em>left</em>), the optimality criterion plot (<em>center</em>), and the regression diagnostic plot (<em>right</em>) for output of function <code>rgrplars()</code> in package <code>robustHD</code>.</figcaption>
+<img src="./inst/doc/paper/figure_rgrplars.svg" style="width:100.0%" alt="Examples of the coefficient plot (left), the optimality criterion plot (center), and the regression diagnostic plot (right) for output of function rgrplars() in package robustHD." /><figcaption aria-hidden="true">Examples of the coefficient plot (<em>left</em>), the optimality criterion plot (<em>center</em>), and the regression diagnostic plot (<em>right</em>) for output of function <code>rgrplars()</code> in package <code>robustHD</code>.</figcaption>
 </figure>
 
 ## Community guidelines
