### 2021-10-01 #3786 @whedon
[REVIEW]: robustHD: An R package for robust regression with high-dimensional data
**Submitting author:** @aalfons (<a href="http://orcid.org/0000-0002-2513-3788">Andreas Alfons</a>)
**Repository:** <a href="https://github.com/aalfons/robustHD/" target ="_blank">https://github.com/aalfons/robustHD/</a>
**Version:** v0.7.1
**Archive:** <a href="https://doi.org/10.25397/eur.16802596.v1" target="_blank">10.25397/eur.16802596.v1</a>
