### 2021-10-01 #3786 @mikldk
@valentint, @msalibian: Thanks for agreeing to review. Please carry out your review in this issue by updating the checklist above and giving feedback in this issue. The reviewer guidelines are available here: <https://joss.readthedocs.io/en/latest/reviewer_guidelines.html>. If possible create issues (and cross-reference) in the submission's repository to avoid too specific discussions in this review thread.

Please note that I asked about the length of the paper in the [pre-review](https://github.com/openjournals/joss-reviews/issues/3779#issuecomment-931275964). Both @aalfons and I would like your input in that. Bear in mind the JOSS guidelines of what the paper should contain: <https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain>.
