### 2021-10-08 #3786 @msalibian
@mikldk I'm happy with the submission (documentation & repository are good to go, as far as I can see). For the paper I have a few suggestions below.

The issue of length is a tricky one: yes, the paper is long, but the two examples illustrate different functionality of the software. One way around this could be to reduce the text output of the examples in the paper (e.g. remove the estimated regression coefficients, etc.) and shorten the description of the data, etc. I think Figure 1 may also go, but I'd keep Figure 2 (the main difference between them being that in Fig 1 one does not get the whole 'path of solutions', because either `sparseLTS` doesn't compute it, or because it's not computed when using prediction error (as opposed to BIC) to select the optimal model, which is something that I'd like @aalfons to clarify, see my comment below).

Specific suggestions:
- Summary: I'd avoid saying "cleaning the data" (lines 18-19 of page 1), specially when advocating for the use of robust estimation methods. Such a phrase opens the door to considering cleaning the data and applying "classical" methods as an alternative "robust" methodology, which is well-known to not work well in general;
- Second paragraph of "Statement of need", rather than referring to the historical value of the package, I'd focus on what it does: implements regularized trimmed-squares regression estimators, the plug-in LARS and its group-wise variant;
- line 57 (page 2), does `sparseLTS` **estimate** the smallest lambda that sets all coefficients to zero, or does it **compute** it?
- line 58, what other options are available for the arguments `mode` and `crit`?
- line 60, presumably the cross-validated prediction error also uses some kind of trimming or robustification, this should be mentioned when discussing how the prediction error is estimated;
- line 66, although the argument `K` of `foldControl` seems fairly self explanatory, I think it'd be helpful to mention what `R` is or does;
- lines 68-79, it'd also be helpful to say something about what is in the "reweighted" column (as compared with the "raw" one);
- line 104, comparing this with the robust lars example, can we also get a full path of solutions with `sparseLTS` if we use `crit = "BIC"`? the different default behaviour is a bit intriguing;
- lines 130-137: how are groups formed? It does not seem to be specified explicitly in the call to `rgrplars` below; is there a default behaviour? can the user override it?
- lines 135-137: I don't think I'm understanding this sentence correctly: "The robust regression estimator used to estimate the submodels uses an initial subsampling of observations(...)" Should "estimate the submodels" be "estimate the regression coefficients corresponding to each submodel" (or something along those lines)? Or is it really trying to determine which groups of features (and thus submodels) to use?
- Also regarding the sentence just above: I seem to remember that the "original" Robust LARS simply run the classical LARS algorithm using robust correlation coefficients, and thus it did not need iterations or initial values, etc. Am I remembering it wrong?
