### 2021-10-12 313ba13
minor change in example for NCI-60 data
