### 2021-10-12 5fd1285
changes in JOSS paper according to reviewer's comments
