+++ inst/doc/paper/paper.md
@@ -20,6 +20,7 @@ affiliations:
   index: 1
 ---
 
+
 # Summary
 
 In regression analysis with high-dimensional data, variable selection is an 
@@ -37,136 +38,35 @@ the popular lasso regression estimator [@tibshirani96].  Selecting the optimal
 model can be done via cross-validation or an information criterion, and various 
 plots are available to illustrate model selection and to evaluate the final 
 model estimates.  Furthermore, the package includes functionality for 
-pre-processing and cleaning the data, such as robust standardization and 
-winsorization.  Finally, `robustHD` follows a clear object-oriented design and 
-takes advantage of C++ code and parallel computing to reduce computing time. 
+pre-processing such as robust standardization and winsorization.  Finally, 
+`robustHD` follows a clear object-oriented design and takes advantage of 
+C++ code and parallel computing to reduce computing time. 
+
 
 # Statement of need
 
-While the authors of @khan07b did provide an R script for robust least angle 
-regression (although the link in their paper appears to be broken), our 
-implementation utilizes C++ code for computational efficiency and provides the 
-convenience of an R package.  Moreover, code for robust groupwise least angle 
-regression and sparse least trimmed squares regression is not available 
-elsewhere.  The R package `robustHD` therefore provides researchers with access 
-to several popular methods for robust regression and variable selection with 
-high-dimensional data.
+While the authors of @khan07b did provide an R script for robust least 
+angle regression (although the link in their paper appears to be broken), 
+the implementation in `robustHD` utilizes C++ code for computational 
+efficiency and provides the convenience of an R package.  Moreover, code 
+for robust groupwise least angle regression and sparse least trimmed 
+squares regression is not available elsewhere.  Package `robustHD` 
+therefore provides researchers with access to several popular methods 
+for robust regression and variable selection with high-dimensional 
+data.  It has been used in many benchmarking studies in the statistical 
+literature [e.g., @chang18; @cohenfreue19; @kurnaz18; @smucler17], as well 
+as in empirical research [e.g., @antczakorlewska21; @stadlbauer20].
 
-At the time of its first release, `robustHD` was to the best of my knowledge 
-the first publicly available R package for robust regression in high 
-dimensions.  In particular, it provided the first generalization of the popular 
-lasso regression estimator that is robust against all types of outliers (that 
-is, including outliers in the predictor space).  Since then, `robustHD` has 
-been used in many benchmarking studies in the statistical literature 
-[e.g., @chang18; @cohenfreue19; @kurnaz18; @smucler17], as well as in empirical 
-research [e.g., @antczakorlewska21; @stadlbauer20].
-
-# Example: Sparse least trimmed squares regression
-
-The well-known NCI-60 cancer cell panel [@reinhold12] is used to illustrate 
-the functionality for sparse least trimmed squares regression. The protein 
-expressions for a specific protein are selected as the response variable, 
-and the gene expressions of the 100 genes that have the highest (robustly 
-estimated) correlations with the response variable are screened as candidate 
-predictors.
-
-```
-# load package and data
-library("robustHD")
-data("nci60")  # contains matrices 'protein' and 'gene'
-
-# define response variable
-y <- protein[, 92]
-# screen most correlated predictor variables
-correlations <- apply(gene, 2, corHuber, y)
-keep <- partialOrder(abs(correlations), 100, decreasing = TRUE)
-X <- gene[, keep]
-```
-
-Sparse least trimmed squares is a regularized estimator of the linear 
-regression model, whose results depend on a non-negative regularization 
[…]
