### 2021-10-12 ba235e5
minor change in JOSS paper
