### 2021-10-12 09cb6e7
minor change in JOSS paper
