### 2021-10-12 0053c2c
minor change in JOSS paper
