### 2021-10-12 f9af757
minor change in JOSS paper
