+++ README.md
@@ -10,23 +10,23 @@ interpretability of the resulting models due to the smaller number of
 variables. However, robust methods are necessary to prevent outlying
 data points from distorting the results. The add-on package `robustHD`
 for the statistical computing environment `R` provides functionality for
-robust linear model selection with high-dimensional data. More
-specifically, the implemented functionality includes robust least angle
-regression ([Khan et al.,
+robust linear regression and model selection with high-dimensional data.
+More specifically, the implemented functionality includes robust least
+angle regression ([Khan et al.,
 2007](https://doi.org/10.1198/016214507000000950)), robust groupwise
 least angle regression ([Alfons et al.,
 2016](https://doi.org/10.1016/j.csda.2015.02.007)), as well as sparse
-least trimmed squares regression [Alfons et al.,
-2013](https://doi.org/10.1214/12-AOAS575). The latter can be seen as a
+least trimmed squares regression ([Alfons et al.,
+2013](https://doi.org/10.1214/12-AOAS575)). The latter can be seen as a
 trimmed version of the popular lasso regression estimator ([Tibshirani,
 1996](https://doi.org/10.1111/j.2517-6161.1996.tb02080.x)). Selecting
 the optimal model can be done via cross-validation or an information
 criterion, and various plots are available to illustrate model selection
 and to evaluate the final model estimates. Furthermore, the package
-includes functionality for pre-processing and cleaning the data, such as
-robust standardization and winsorization. Finally, `robustHD` follows a
-clear object-oriented design and takes advantage of C++ code and
-parallel computing to reduce computing time.
+includes functionality for pre-processing such as robust standardization
+and winsorization. Finally, `robustHD` follows a clear object-oriented
+design and takes advantage of C++ code and parallel computing to reduce
+computing time.
 
 ## Main functionality
 
@@ -42,10 +42,11 @@ parallel computing to reduce computing time.
 
 -   `corHuber()`: Robust correlation based on winsorization.
 
--   `winsorize()`: Data cleaning by winsorization.
+-   `winsorize()`: Winsorization of the data.
 
--   `robStandardize()`: Data standardization with given functions for
-    computing center and scale. By default, the median and MAD are used.
+-   `robStandardize()`: Robust standardization of the data with given
+    functions for computing center and scale. By default, the median and
+    MAD are used.
 
 ## Installation
 
@@ -99,16 +100,18 @@ regression model, whose results depend on a non-negative regularization
 parameter \[see [Alfons et al.,
 2013](https://doi.org/10.1214/12-AOAS575)\]. In general, a larger value
 of this regularization parameter yields more regression coefficients
-being set to zero, which can be seen as a form of variable selection for
-the linear model.
+being set to zero, which can be seen as a form of variable selection.
 
 For convenience, `sparseLTS()` can internally estimate the smallest
 value of the regularization parameter that sets all coefficients to
 zero. With `mode = "fraction"`, the values supplied via the argument
 `lambda` are then taken as fractions of this estimated value (i.e., they
 are multiplied with the internally estimated value). In this example,
-the prediction error is estimated via 5-fold cross-validation for
-selecting the optimal value of the regularization parameter, and the
+the optimal value of the the regularization parameter is selected by
+estimating the prediction error (`crit = "PE"`) via 5-fold
+cross-validation with one replication
+(`splits = foldControl(K = 5, R = 1)`). The default prediction loss
+function is the root trimmed mean squared prediction error. Finally, the
 seed of the random number generator is supplied for reproducibility.
 
 ``` text
@@ -154,16 +157,16 @@ fit
     ## Residual scale estimate: 0.62751742
 
 Among other information, the output prints the results of the final
-model fit, which here consists of 17 genes (with non-zero coefficients).
[…]
