### 2021-10-12 #3786 @aalfons
@mikldk @msalibian @valentint - I created a revised version of the manuscript which is reduced in length and fits on 5 pages. I'll give detailed responses to each reviewer in separate comments below.
