### 2021-10-12 #3786 @aalfons
Reply to @msalibian:

Thanks again for your helpful suggestions on reducing the length. I decided to keep the output on the regression coefficients, but for `sparseLTS()` I excluded the output regarding the cross-validation.  This also avoided a discussion on the raw and reweighted estimators (see also my reply to one of the comments below). Moreover, I followed your suggestion and removed Figure 1 but kept Figure 2.  With this change, it seemed more appropriate to present the `rgrplars()` example first, as it is more detailed.  Furthermore, I think that in this case the length of the paper can be further reduced by avoiding the discussion on when the coefficient plot is available (see my reply to one of the comments below for details).

ad 1) (avoid saying "cleaning the data") -> Done (see page 1, line 19).

ad 2) (shortening the Statement of Need) -> As suggested, I removed two sentences on the historical value of the package, but I kept the sentence referring to the use of `robustHD` in the literature, as it underlines its usefulness (page 1, lines 29-32). But I'd be happy for some direction by @mikldk on this (see also my reply below to a related point by @valentint).

ad 3) (**estimate** or **compute** smallest lambda that sets everything to zero) -> As clarified this in my earlier reply, the word "estimate" is correct here.

ad 4) (other options for the arguments `mode` and `crit`) -> In the spirit of not (unnecessarily) extending the length of the paper, I think it is better not to include a discussion on other options for the arguments `mode` and `crit`.  Interested readers can find this information in the help file.

ad 5) (cross-validated prediction error) -> I added a sentence that for sparse LTS, the default prediction loss function
is the root trimmed mean squared prediction error (page 3, lines 114-115).

ad 6) (clarification of arguments of `foldControl()`) -> I took greater care to explain the different arguments in the function call, including an explanation that `splits = foldControl(K = 5, R = 1)` defines 5-fold cross-validation with one replication (page 3, lines 112-114).

ad 7) (clarifying "reweighted" and "raw" columns of output) -> To shorten the paper, I followed your suggestion to remove some of the text output in the `sparseLTS()` example.  The cross-validation results are no longer shown, avoiding a discussion on the "raw" and "reweighted" columns altogether.

ad 8) (whether we can get a full path of `sparseLTS()` solutions with `crit = "BIC"`) -> As clarified in my earlier reply, this behavior has to do with how the final model is selected, irrespective of the modeling technique.  With switching the order of the examples and removing the plots for the `sparseLTS()` example (see the discussion above), I found it better to avoid this discussion altogether to shorten the paper. Regarding plots of results from `sparseLTS()`, I now simply write: "Similar plots as in Figure 1 are available to visualize the results." (page 3, lines 140-141)  This is a bit of an oversimplification, as the coefficient plot is not available for this specific example, but I think that this simplification acceptable since the plot is indeed available if `crit = "BIC"` is used in `sparseLTS()`.

ad 9) (how groups are formed in `rgrplars()`) -> I rewrote the corresponding paragraph.  Specifically, I now write: "Through the formula interface, function `rgrplars()` by default takes each categorical variable (`factor`) as a group of dummy variables while all remaining variables are taken individually.  However, the group assignment can be defined by the user through argument `assign`." (page 2, lines 53-56)

ad 10) (confusing sentence regarding the robust regression estimator in `rgrplars()`) -> In line with the clarification from my earlier reply, I now write: "Note that each submodel along the sequence is fitted using a robust regression estimator with a non-deterministic algorithm, hence the seed of the random number generator is supplied for reproducibility." (page 2, lines 58-60)

ad 11) (clarification on robust LARS) -> As already clarified in my earlier reply, also robust LARS uses a two-step strategy where first the variables are sequences, and then submodel along the sequence are estimated with robust regressions.

Thanks again for the helpful comments, and please let me know if anything is still unclear.
