### 2021-10-12 #3786 @valentint
@mikldk First of all I would like to recommend the revised paper for publishing.
The technical issues I had mentioned before could be discussed out of this review.

@aalfons Installation: yes, it seems that it is my local problem, but it occurs on all my PCs :) and I still cannot find a way out of it. Since I have had the issue with other R packages, it is clear that it does not concern robustHD.

@aalfons @msalibian The timing: I got curious now :) I am on a relatively new Surface Book, i7 processor, and 8GB RAM. Please find below the script with my approximate timings and let me know if I am doing something wrong.

library("robustHD")
library(tictoc)

data("nci60")

y <- protein[, 92]
correlations <- apply(gene, 2, corHuber, y)
keep <- partialOrder(abs(correlations), 100, decreasing = TRUE)
X <- gene[, keep]

## First I run sparseLTS() with the  sequence of 10 lamdas, as in the example on GitHub - it takes approx 5 minutes.
lambda <- seq(0.01, 0.5, length.out=10)
tic()
fit <- sparseLTS(X, y, lambda=lambda, mode = "fraction", crit = "PE", splits = foldControl(K = 5, R = 1), seed = 20210507)
toc()

## 2. Then I run three times sparseLTS() for three different values of lambda - it is quite fast, less then 10 seconds each
tic()
fit <- sparseLTS(X, y, lambda=0.01, mode = "fraction", crit = "PE", splits = foldControl(K = 5, R = 1), seed = 20210507)
fit <- sparseLTS(X, y, lambda=0.255, mode = "fraction", crit = "PE", splits = foldControl(K = 5, R = 1), seed = 20210507)
fit <- sparseLTS(X, y, lambda=0.05, mode = "fraction", crit = "PE", splits = foldControl(K = 5, R = 1), seed = 20210507)
toc()

## 2. And finally I run sparseLTS() with a grid containing the same three lambdas - the elapsed time is more than five times larger
lambda <- c(0.01, 0.255, 0.5)
tic()
fit <- sparseLTS(X, y, lambda=lambda, mode = "fraction", crit = "PE", splits = foldControl(K = 5, R = 1), seed = 20210507)
toc()
