### 2021-10-12 #3786 @msalibian
@valentint This is intriguing. Below I copy my Surface system information, and here's the output of the first part of your script. I don't know what may explain the difference.

```
> library("robustHD")
Loading required package: ggplot2
Loading required package: perry
Loading required package: parallel
Loading required package: robustbase
Warning messages:
1: package ‘perry’ was built under R version 4.1.1
2: package ‘robustbase’ was built under R version 4.1.1
> library(tictoc)
Warning message:
package ‘tictoc’ was built under R version 4.1.1
> data("nci60")
> y <- protein[, 92]
> correlations <- apply(gene, 2, corHuber, y)
> keep <- partialOrder(abs(correlations), 100, decreasing = TRUE)
> X <- gene[, keep]
> lambda <- seq(0.01, 0.5, length.out=10)
> tic()
> fit <- sparseLTS(X, y, lambda=lambda, mode = "fraction", crit = "PE", splits = foldControl(K = 5, R = 1), seed = 20210507)
> toc()
28.5 sec elapsed
>
> system.time(
+   fit <- sparseLTS(X, y, lambda=lambda, mode = "fraction", crit = "PE", splits = foldControl(K = 5, R = 1), seed = 20210507)
+ )
   user  system elapsed
  26.91    0.01   27.98
> sessionInfo()
R version 4.1.0 (2021-05-18)
Platform: x86_64-w64-mingw32/x64 (64-bit)
Running under: Windows 10 x64 (build 19043)

Matrix products: default

locale:
[1] LC_COLLATE=English_Canada.1252  LC_CTYPE=English_Canada.1252
[3] LC_MONETARY=English_Canada.1252 LC_NUMERIC=C
[5] LC_TIME=English_Canada.1252

attached base packages:
[1] parallel  stats     graphics  grDevices utils     datasets  methods   base

other attached packages:
[1] tictoc_1.0.1      robustHD_0.7.1    robustbase_0.93-9 perry_0.3.0       ggplot2_3.3.5

 [1] Rcpp_1.0.7       magrittr_2.0.1   MASS_7.3-54      tidyselect_1.1.1 munsell_0.5.0
 [6] colorspace_2.0-2 R6_2.5.1         rlang_0.4.11     fansi_0.5.0      dplyr_1.0.7
[11] tools_4.1.0      grid_4.1.0       gtable_0.3.0     utf8_1.2.2       withr_2.4.2
[16] ellipsis_0.3.2   tibble_3.1.5     lifecycle_1.0.1  crayon_1.4.1     purrr_0.3.4
[21] vctrs_0.3.8      glue_1.4.2       compiler_4.1.0   DEoptimR_1.0-9   pillar_1.6.3
[26] generics_0.1.0   scales_1.1.1     pkgconfig_2.0.3
```

![image](https://user-images.githubusercontent.com/5422467/137037223-ed340322-5071-4082-9cd1-52e60ed245e6.png)
