### 2021-10-13 92bb3f7
fixed documentation of sparseLTS() for the case of only one value of the penalty parameter
