+++ DESCRIPTION
@@ -2,7 +2,7 @@ Package: robustHD
 Type: Package
 Title: Robust Methods for High-Dimensional Data
 Version: 0.7.1
-Date: 2021-09-29
+Date: 2021-10-13
 Depends: 
     R (>= 3.5.0), 
     ggplot2 (>= 0.9.2), 
