### 2021-10-13 2ed070c
minor language fix in JOSS paper and updated package version in corresponding bib file
