+++ inst/doc/paper/paper.bib
@@ -2,7 +2,7 @@
   title = {{robustHD}: Robust Methods for High-Dimensional Data},
   author = {Alfons, A.},
   year = {2021},
-  note = {R package version~0.7.0},
+  note = {R package version~0.7.1},
   url = {https://github.com/aalfons/robustHD/}
 }
 
