### 2021-11-03 #3786 @aalfons
@kthyng - The metadata on the archive is now publicly visible, including the sentence regarding the version number.

It turns out that the change in the metadata didn't trigger a new DOI after all, the current DOI is still correct. So everything should now be ready to go.
