### 2021-11-04 #3786 @aalfons
Thanks @kthyng - the DOI seems to work!

And thanks again @mikldk, @valentint, and @msalibian for the constructive comments.
